change auth provider page
